requirements.txt
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.1
asyncpg==0.25.0
black==22.3.0
cffi==1.15.0
click==8.1.3
databases==0.5.5
fastapi==0.77.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycparser==2.21
pydantic==1.9.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.2.0
uvicorn==0.17.6
